+++ sites/partners/src/listings/PaperListingForm/index.tsx
@@ -22,7 +22,6 @@ import {
 import { useForm, FormProvider } from "react-hook-form"
 import {
   ListingStatus,
-  CSVFormattingType,
   ListingApplicationAddressType,
   Unit,
   Listing,
@@ -151,7 +150,6 @@ const defaults: FormListing = {
   costsNotIncluded: "",
   creditHistory: "",
   criminalBackground: "",
-  CSVFormattingType: CSVFormattingType.basic,
   depositMax: "0",
   depositMin: "0",
   disableUnitsAccordion: false,
@@ -525,6 +523,8 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
           removeEmptyFields(formattedData, [
             "applicationPickUpAddressType",
             "applicationDropOffAddressType",
+            "applicationDueDate",
+            "applicationDueTime",
           ])
           const result = editMode
             ? await listingsService.update({